backend/requirements.txt
# Backend API Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlite3  # Built-in with Python
# For production: psycopg2-binary==2.9.9  # PostgreSQL
# For production: asyncpg==0.29.0  # Async PostgreSQL

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Caching
redis==5.0.1

# Logging
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Existing dependencies from main project
python-dotenv
qdrant-client
sentence-transformers
requests


